pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
name = "ddt4all"
version = "3.0.7.1"
description = "DDT4All is a comprehensive tool to create your own ECU parameters screens and connect to a CAN network with various OBD-II interfaces including ELM327, Vlinker FS, VGate, ObdLink SX, and ELS27 adapters."
requires-python = ">=3.8"

dependencies = [
    "PyQt5>=5.15.0,<5.16.0",
    "PyQtWebEngine>=5.15.0,<5.16.0",

    "pyserial==3.5",
    "pyusb==1.2.1",
    "crcmod==1.7",

    # Windows only
    "pywin32>=227; sys_platform == 'win32'",

    # Safe ranges to avoid breaking changes
    "deep_translator>=1.11.4,<2.0.0",
    "polib>=1.2.0,<2.0.0",
]

[project.optional-dependencies]

dev = [
    "PyQt5-stubs>=5.15.0,<5.16.0",

    # pytest versions that still support Python 3.8
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-mock>=3.14.1",
    "pytest-qt",
]

can = [
    "python-can>=4.0.0,<5.0.0",
    "can-isotp>=2.0.0,<3.0.0",
    "obd>=0.7.1,<1.0.0",
]

network = [
    "requests>=2.28.0,<3.0.0",
    "websockets>=10.0,<12.0.0",
]

bluetooth = [
    # Linux / Windows
    "pybluez-updated>=0.31; sys_platform != 'darwin'",

    # macOS fallback (opcional mas recomendado)
    "bleak>=0.20.0; sys_platform == 'darwin'",
]

[project.scripts]
ddt4all = "ddt4all.main:main"

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.coverage.run]
source = ["ddt4all"]
branch = true

[tool.ruff]
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "UP", "FA"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests/integration", "tests/unit", "tests/smoke"]
addopts = "-q"